+++ nym-api/src/support/cli/run.rs
@@ -20,10 +20,10 @@ use crate::status::{ApiStatusState, SignerState};
 use crate::support::caching::cache::SharedCache;
 use crate::support::config::helpers::try_load_current_config;
 use crate::support::config::{Config, DEFAULT_CHAIN_STATUS_CACHE_TTL};
-use crate::support::http::state::{
-    AppState, ChainStatusCache, ForcedRefresh, ShutdownHandles, TASK_MANAGER_TIMEOUT_S,
-};
-use crate::support::http::RouterBuilder;
+use crate::support::http::state::chain_status::ChainStatusCache;
+use crate::support::http::state::force_refresh::ForcedRefresh;
+use crate::support::http::state::AppState;
+use crate::support::http::{RouterBuilder, ShutdownHandles, TASK_MANAGER_TIMEOUT_S};
 use crate::support::nyxd;
 use crate::support::storage::runtime_migrations::m001_directory_services_v2_1::migrate_to_directory_services_v2_1;
 use crate::support::storage::NymApiStorage;
